+++ app/api/tags/relation/route.ts
@@ -3,14 +3,46 @@ import { prisma } from "@/app/lib/config/prisma";
 import { StatusCodes } from "http-status-codes";
 
 export async function POST(request: Request) {
+  let parentId: number;
+  let childId: number;
   try {
-    const { parentId, childId } = await request.json();
-    console.debug("Received data:", { parentId, childId });
+    const json = await request.json();
+    parentId = json.parentId;
+    childId = json.childId;
 
-    if (parentId === childId) {
+    console.debug("Received data:", { parentId, childId });
+  } catch (error) {
+    return NextResponse.json(
+      { error: `Error parsing request body, ${error}` },
+      { status: StatusCodes.BAD_REQUEST }
+    );
+  }
+
+  if (parentId === childId) {
+    return NextResponse.json(
+      { error: "Parent cannot be equal to Child" },
+      { status: StatusCodes.BAD_REQUEST }
+    );
+  }
+
+  try {
+    const parentTag = await prisma.tag.findUnique({
+      where: { id: parentId },
+    });
+    if (parentTag === null) {
       return NextResponse.json(
-        { error: "Parent cannot be equal to Child" },
-        { status: StatusCodes.BAD_REQUEST }
+        { error: "Parent tag not found" },
+        { status: StatusCodes.NOT_FOUND }
+      );
+    }
+
+    const childTag = await prisma.tag.findUnique({
+      where: { id: childId },
+    });
+    if (childTag === null) {
+      return NextResponse.json(
+        { error: "Child tag not found" },
+        { status: StatusCodes.NOT_FOUND }
       );
     }
 
@@ -28,32 +60,6 @@ export async function POST(request: Request) {
   }
 }
 
-export async function PATCH(request: Request) {
-  try {
-    const { id, parentId, childId } = await request.json();
-
-    if (parentId === childId) {
-      return NextResponse.json(
-        { error: "Parent cannot be equal to Child" },
-        { status: StatusCodes.BAD_REQUEST }
-      );
-    }
-
-    const updated = await prisma.tagRelation.update({
-      where: { id },
-      data: { parentId, childId },
-    });
-
-    return NextResponse.json(updated);
-  } catch (error) {
-    console.error("Error update tag relation:", error);
-    return NextResponse.json(
-      { error: "Error updating tag relation" },
-      { status: StatusCodes.INTERNAL_SERVER_ERROR }
-    );
-  }
-}
-
 export async function GET() {
   try {
     const relations = await prisma.tagRelation.findMany();
@@ -62,31 +68,6 @@ export async function GET() {
     console.error("Error fetching tag relations:", error);
     return NextResponse.json(
       { error: "Error fetching tag relations" },
-      { status: StatusCodes.INTERNAL_SERVER_ERROR }
-    );
-  }
-}
-
-export async function DELETE(request: Request) {
-  try {
-    const { id } = await request.json();
-
-    if (!id) {
-      return NextResponse.json(
-        { error: "ID is required" },
-        { status: StatusCodes.BAD_REQUEST }
-      );
-    }
-
-    const deleted = await prisma.tagRelation.delete({
-      where: { id },
-    });
-
-    return NextResponse.json(deleted);
-  } catch (error) {
-    console.error("Error deleting tag:", error);
-    return NextResponse.json(
-      { error: "Error deleting tag" },
       { status: StatusCodes.INTERNAL_SERVER_ERROR }
     );
   }